rename: 기타 컴포넌트 옮기기
Related to: #32
